pyproject.toml
[tool.poetry]
name = "weatherapp"
version = "0.1.0"
description = ""
authors = ["Anton Skvortsov <askvrtsv@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
django-debug-toolbar = "^4.3.0"
djangorestframework-dataclasses = "^1.3.1"
requests = "^2.31.0"
pytelegrambotapi = "^4.15.4"
cachetools = "^5.3.2"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.31.0.20240125"
types-cachetools = "^5.3.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
